LPS-88626 Fragments Editor constants moved to constants file
