+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/components/fragment_entry_link/FragmentEntryLinkList.es.js
@@ -21,7 +21,7 @@ import {
 import {
 	FRAGMENTS_EDITOR_ITEM_BORDERS,
 	FRAGMENTS_EDITOR_ITEM_TYPES
-} from '../../reducers/placeholders.es';
+} from '../../utils/constants';
 import {
 	getFragmentColumn,
 	getItemMoveDirection,
