+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/components/sidebar/layouts/SidebarLayoutsPanel.es.js
@@ -7,7 +7,7 @@ import {
 	CLEAR_DROP_TARGET,
 	UPDATE_DROP_TARGET
 } from '../../../actions/actions.es';
-import {FRAGMENTS_EDITOR_ITEM_TYPES} from '../../../reducers/placeholders.es';
+import {FRAGMENTS_EDITOR_ITEM_TYPES} from '../../../utils/constants';
 import SidebarLayoutsDragDrop from './utils/SidebarLayoutsDragDrop.es';
 import {Store} from '../../../store/store.es';
 import templates from './SidebarLayoutsPanel.soy';
